add base exception
