+++ src/common/error/src/error.rs
@@ -48,13 +48,6 @@ impl From<arrow2::error::Error> for DaftError {
     }
 }
 
-#[cfg(feature = "python")]
-impl From<pyo3::PyErr> for DaftError {
-    fn from(error: pyo3::PyErr) -> Self {
-        DaftError::PyO3Error(error)
-    }
-}
-
 impl From<serde_json::Error> for DaftError {
     fn from(error: serde_json::Error) -> Self {
         DaftError::IoError(error.into())
@@ -70,20 +63,6 @@ impl From<io::Error> for DaftError {
 impl From<regex::Error> for DaftError {
     fn from(error: regex::Error) -> Self {
         DaftError::ValueError(error.to_string())
-    }
-}
-
-#[cfg(feature = "python")]
-impl std::convert::From<DaftError> for pyo3::PyErr {
-    fn from(err: DaftError) -> pyo3::PyErr {
-        use pyo3::exceptions::{PyFileNotFoundError, PyValueError};
-        match err {
-            DaftError::PyO3Error(pyerr) => pyerr,
-            DaftError::FileNotFound { path, source } => {
-                PyFileNotFoundError::new_err(format!("File: {path} not found\n{source}"))
-            }
-            _ => PyValueError::new_err(err.to_string()),
-        }
     }
 }
