+++ src/common/error/src/lib.rs
@@ -1,3 +1,5 @@
 mod error;
 pub use error::DaftError;
 pub use error::DaftResult;
+#[cfg(feature = "python")]
+mod python;
